Started messing around. Mostly basic controls
Steering the train and the train always go forward (depending on the rotation of the train of course). It also actually act like a train, so there isn't something called Inertia when the train turns and doesn't end up drifting (which was stupid). Note to self : check if the train speed can change while chugging along (it should still have inertia from going forward)